Move MvcModule back to base api.
