+++ src/Microsoft.Health.Fhir.Api/Features/Routing/ResourceTypesRouteConstraint.cs
@@ -3,23 +3,15 @@
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
-using System;
 using EnsureThat;
-using Hl7.Fhir.Model;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
+using Microsoft.Health.Fhir.Core.Models;
 
 namespace Microsoft.Health.Fhir.Api.Features.Routing
 {
-    public class CompartmentResourceTypesRouteConstraint : IRouteConstraint
+    public class ResourceTypesRouteConstraint : IRouteConstraint
     {
-        private readonly ResourceTypesRouteConstraint _resourceTypesRouteConstraint;
-
-        public CompartmentResourceTypesRouteConstraint()
-        {
-            _resourceTypesRouteConstraint = new ResourceTypesRouteConstraint();
-        }
-
         public bool Match(HttpContext httpContext, IRouter route, string routeKey, RouteValueDictionary values, RouteDirection routeDirection)
         {
             EnsureArg.IsNotNull(httpContext, nameof(httpContext));
@@ -27,17 +19,9 @@ namespace Microsoft.Health.Fhir.Api.Features.Routing
             EnsureArg.IsNotNullOrEmpty(routeKey, nameof(routeKey));
             EnsureArg.IsNotNull(values, nameof(values));
 
-            if (values.TryGetValue(KnownActionParameterNames.ResourceType, out var resourceTypeObj) &&
-                resourceTypeObj is string resourceType &&
-                !string.IsNullOrEmpty(resourceType))
+            if (values.TryGetValue(KnownActionParameterNames.ResourceType, out var resourceTypeObj) && resourceTypeObj is string resourceType && !string.IsNullOrEmpty(resourceType))
             {
-                // Don't validate wild card.
-                if (resourceType == "*")
-                {
-                    return true;
-                }
-
-                return _resourceTypesRouteConstraint.Match(httpContext, route, routeKey, values, routeDirection);
+                return ModelInfoProvider.IsKnownResource(resourceType);
             }
 
             return false;
