+++ src/Microsoft.Health.Fhir.Core/Models/IModelInfoProvider.cs
@@ -15,6 +15,8 @@ namespace Microsoft.Health.Fhir.Core.Models
 
         string GetFhirTypeNameForType(Type type);
 
+        bool IsKnownCompartment(string compartmentType);
+
         bool IsKnownResource(string name);
 
         bool IsKnownCompartmentType(string compartmentType);
